requirements.txt
flask
flask-swagger-ui
cassandra-driver
